Topology: surface the network map as a third tab on the Topology page
The Topology page promised "a map of your system" while silently
omitting the physical layer. It now has Service Map | Infrastructure |
Network tabs. The network map is surfaced, not moved: the LLDP
fetch/render/device-navigation logic is extracted into a shared
NetworkTopologyView component rendered from both the Topology page and
the Network Devices section, where network engineers still expect it.

The network tab is a live view, so the telemetry time-range picker and
truncation banner hide while it is active (with the hint text switching
to say so). The graph's empty state now carries a "set up network
device monitoring" link — injected by the view rather than imported
into the graph component, whose pure layout export is unit-tested in a
plain node environment and must stay free of router/window imports.
